Move check for rst error. Rename MockGapicLayer.
